Move ExtractCommonNameAndDNSNames to pkg/shared
