+++ pkg/shared/utils_certificate.go
@@ -1,0 +1,39 @@
+/*
+ * SPDX-FileCopyrightText: 2019 SAP SE or an SAP affiliate company and Gardener contributors
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package shared
+
+import (
+	"crypto/x509"
+	"encoding/pem"
+	"fmt"
+)
+
+// ExtractCommonNameAnDNSNames extracts values from a CSR (Certificate Signing Request).
+func ExtractCommonNameAnDNSNames(csr []byte) (cn *string, san []string, err error) {
+	certificateRequest, err := extractCertificateRequest(csr)
+	if err != nil {
+		err = fmt.Errorf("parsing CSR failed: %w", err)
+		return
+	}
+	cnvalue := certificateRequest.Subject.CommonName
+	if cnvalue != "" {
+		cn = &cnvalue
+	}
+	san = certificateRequest.DNSNames[:]
+	for _, ip := range certificateRequest.IPAddresses {
+		san = append(san, ip.String())
+	}
+	return
+}
+
+func extractCertificateRequest(csr []byte) (*x509.CertificateRequest, error) {
+	block, _ := pem.Decode(csr)
+	if block == nil {
+		return nil, fmt.Errorf("decoding CSR failed")
+	}
+	return x509.ParseCertificateRequest(block.Bytes)
+}
